Updated UI and fixed product card layout

+++ src/app/Shared/Components/features/Category/category-list/category-list.ts
@@ -9,25 +9,21 @@ import { RouterLink } from "@angular/router";
   styleUrl: './category-list.css',
 })
 export class CategoryList {
-  private categoryService = inject(CategoryService);
+private categoryService = inject(CategoryService);
 
-  private getAllCategoryRef = this.categoryService.getAllCategory();
+  private categoryRef = this.categoryService.getAllCategory();
 
-  isLoading = this.getAllCategoryRef.isLoading;
-  isError = this.getAllCategoryRef.error;
-  categories = this.getAllCategoryRef.value;
+  isLoading = this.categoryRef.isLoading;
+  isError = this.categoryRef.error;
+  categories = this.categoryRef.value;
 
-
-  deleteCategory(id : string)
-  {
-    this.categoryService.deleteCategory(id).subscribe({
-      next  : ()=> {
-        window.location.reload();
-      },
-      error : () => {
-        console.error("Something went wrong")
-      }
-  
-    })
+  deleteCategory(id: string) {
+    if (confirm('Are you sure you want to delete this category?')) {
+      this.categoryService.deleteCategory(id).subscribe({
+        next: () => {
+          console.log("Deleted successfully");
+        }
+      });
+    }
   }
 }
